Move github types out of big types folder
