+++ src/blocks/jobStatuses.ts
@@ -1,4 +1,4 @@
-import { Annotation } from '@/types';
+import { Annotation } from '@/types/github';
 
 /**
  * Transform GitHub Markdown link to Slack link
